+++ internal/platform-support/tegra/csv_test.go
@@ -90,10 +90,9 @@ func TestDiscovererFromCSVFiles(t *testing.T) {
 			expectedHooks: []discover.Hook{
 				{
 					Lifecycle: "createContainer",
-					Path:      "/usr/bin/nvidia-ctk",
+					Path:      "/usr/bin/nvidia-cdi-hook",
 					Args: []string{
-						"nvidia-ctk",
-						"hook",
+						"nvidia-cdi-hook",
 						"create-symlinks",
 						"--link",
 						"/usr/lib/aarch64-linux-gnu/tegra/libv4l2_nvargus.so::/usr/lib/aarch64-linux-gnu/libv4l/plugins/nv/libv4l2_nvargus.so",
@@ -147,10 +146,9 @@ func TestDiscovererFromCSVFiles(t *testing.T) {
 			expectedHooks: []discover.Hook{
 				{
 					Lifecycle: "createContainer",
-					Path:      "/usr/bin/nvidia-ctk",
+					Path:      "/usr/bin/nvidia-cdi-hook",
 					Args: []string{
-						"nvidia-ctk",
-						"hook",
+						"nvidia-cdi-hook",
 						"create-symlinks",
 						"--link",
 						"/usr/lib/aarch64-linux-gnu/tegra/libv4l2_nvargus.so::/usr/lib/aarch64-linux-gnu/libv4l/plugins/nv/libv4l2_nvargus.so",
@@ -189,7 +187,7 @@ func TestDiscovererFromCSVFiles(t *testing.T) {
 
 			o := tegraOptions{
 				logger:              logger,
-				nvidiaCTKPath:       "/usr/bin/nvidia-ctk",
+				nvidiaCDIHookPath:   "/usr/bin/nvidia-cdi-hook",
 				csvFiles:            []string{"dummy"},
 				ignorePatterns:      tc.ignorePatterns,
 				symlinkLocator:      tc.symlinkLocator,