+++ pkg/nvcdi/driver-wsl_test.go
@@ -92,8 +92,8 @@ func TestNvidiaSMISymlinkHook(t *testing.T) {
 			expectedHooks: []discover.Hook{
 				{
 					Lifecycle: "createContainer",
-					Path:      "nvidia-ctk",
-					Args: []string{"nvidia-ctk", "hook", "create-symlinks",
+					Path:      "nvidia-cdi-hook",
+					Args: []string{"nvidia-cdi-hook", "create-symlinks",
 						"--link", "nvidia-smi::/usr/bin/nvidia-smi"},
 				},
 			},
@@ -112,8 +112,8 @@ func TestNvidiaSMISymlinkHook(t *testing.T) {
 			expectedHooks: []discover.Hook{
 				{
 					Lifecycle: "createContainer",
-					Path:      "nvidia-ctk",
-					Args: []string{"nvidia-ctk", "hook", "create-symlinks",
+					Path:      "nvidia-cdi-hook",
+					Args: []string{"nvidia-cdi-hook", "create-symlinks",
 						"--link", "/some/path/nvidia-smi::/usr/bin/nvidia-smi"},
 				},
 			},
@@ -132,8 +132,8 @@ func TestNvidiaSMISymlinkHook(t *testing.T) {
 			expectedHooks: []discover.Hook{
 				{
 					Lifecycle: "createContainer",
-					Path:      "nvidia-ctk",
-					Args: []string{"nvidia-ctk", "hook", "create-symlinks",
+					Path:      "nvidia-cdi-hook",
+					Args: []string{"nvidia-cdi-hook", "create-symlinks",
 						"--link", "/some/path/nvidia-smi::/usr/bin/nvidia-smi"},
 				},
 			},
@@ -143,9 +143,9 @@ func TestNvidiaSMISymlinkHook(t *testing.T) {
 	for _, tc := range testCases {
 		t.Run(tc.description, func(t *testing.T) {
 			m := nvidiaSMISimlinkHook{
-				logger:        logger,
-				mountsFrom:    tc.mounts,
-				nvidiaCTKPath: "nvidia-ctk",
+				logger:            logger,
+				mountsFrom:        tc.mounts,
+				nvidiaCDIHookPath: "nvidia-cdi-hook",
 			}
 
 			devices, err := m.Devices()